+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 master
 ===
 
+* Add `resources` class method to extensions to allow simple string-based resource generation.
 * rename `app.add_to_instance` to `Extension.expose_to_application` for adding extension-local methods to the shared app instance.
 * rename `app.add_to_config_context` to `Extension.expose_to_config` for adding extension-local methods to the sandboxed scope of `config.rb`
 * Add `Extension.expose_to_templates`, which auto binds copies of extension-local methods into a Template context.